--- apps/storybook/src/stories/useTable.stories.tsx
@@ -1,352 +1,0 @@
-import {
-  Table,
-  TableBody,
-  TableCaption,
-  TableCell,
-  TableFooter,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "../components/table.tsx";
-import { useTable } from "@arkejs/table";
-import {
-  Pagination,
-  PaginationContent,
-  PaginationEllipsis,
-  PaginationItem,
-  PaginationLink,
-  PaginationNext,
-  PaginationPrevious,
-} from "../components/pagination.tsx";
-
-export default {
-  title: "useTable",
-};
-
-const invoices = [
-  {
-    invoice: "INV001",
-    paymentStatus: "Paid",
-    totalAmount: "$250.00",
-    paymentMethod: "Credit Card",
-  },
-  {
-    invoice: "INV002",
-    paymentStatus: "Pending",
-    totalAmount: "$150.00",
-    paymentMethod: "PayPal",
-  },
-  {
-    invoice: "INV003",
-    paymentStatus: "Unpaid",
-    totalAmount: "$350.00",
-    paymentMethod: "Bank Transfer",
-  },
-  {
-    invoice: "INV004",
-    paymentStatus: "Paid",
-    totalAmount: "$450.00",
-    paymentMethod: "Credit Card",
-  },
-  {
-    invoice: "INV005",
-    paymentStatus: "Paid",
-    totalAmount: "$550.00",
-    paymentMethod: "PayPal",
-  },
-  {
-    invoice: "INV006",
-    paymentStatus: "Pending",
-    totalAmount: "$200.00",
-    paymentMethod: "Bank Transfer",
-  },
-  {
-    invoice: "INV007",
-    paymentStatus: "Unpaid",
-    totalAmount: "$300.00",
-    paymentMethod: "Credit Card",
-  },
-];
-
-const columns = [
-  {
-    id: "invoice",
-  },
-  {
-    id: "paymentStatus",
-  },
-  {
-    id: "paymentMethod",
-  },
-  {
-    id: "totalAmount",
-  },
-];
-
-export const Default = () => {
-  const table = useTable({ columns, data: invoices });
-
-  return (
-    <Table>
-      <TableCaption>A list of your recent invoices.</TableCaption>
-      <TableHeader>
-        <TableRow>
-          {table.getAllColumns().map((column) => (
-            <TableHead key={column.id}>{column.id}</TableHead>
-          ))}
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {table.getRows().map((row) => (
-          <TableRow key={row.id}>
-            {row.getAllCells().map((cell) => (
-              <TableCell className="font-medium">{cell.getValue()}</TableCell>
-            ))}
-          </TableRow>
-        ))}
-      </TableBody>
-      <TableFooter>
-        <TableRow>
-          <TableCell colSpan={3}>Total</TableCell>
-          <TableCell className="text-right">$2,500.00</TableCell>
-        </TableRow>
-      </TableFooter>
-    </Table>
-  );
-};
-
-export const WithPagination = () => {
-  const table = useTable({ columns, data: invoices });
-  const {
-    pagination: { pageIndex },
-  } = table.getState();
-
-  return (
-    <>
-      <Table>
-        <TableHeader>
-          <TableRow>
-            <TableHead className="w-[100px]">Invoice</TableHead>
-            <TableHead>Status</TableHead>
-            <TableHead>Method</TableHead>
-            <TableHead className="text-right">Amount</TableHead>
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {invoices.map((invoice) => (
-            <TableRow key={invoice.invoice}>
-              <TableCell className="font-medium">{invoice.invoice}</TableCell>
-              <TableCell>{invoice.paymentStatus}</TableCell>
-              <TableCell>{invoice.paymentMethod}</TableCell>
-              <TableCell className="text-right">
-                {invoice.totalAmount}
-              </TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
-        <TableFooter>
-          <TableRow>
-            <TableCell colSpan={3}>Total</TableCell>
-            <TableCell className="text-right">$2,500.00</TableCell>
-          </TableRow>
-        </TableFooter>
-      </Table>
-      <Pagination>
-        <PaginationContent>
-          <PaginationItem>
-            <PaginationPrevious onClick={() => null} />
-          </PaginationItem>
-          <PaginationItem>
-            <PaginationLink
-              onClick={() => table.setPageIndex(0)}
-              isActive={pageIndex === 0}
-            >
-              1
-            </PaginationLink>
-          </PaginationItem>
-          <PaginationItem>
-            <PaginationLink
-              onClick={() => table.setPageIndex(1)}
-              isActive={pageIndex === 1}
-            >
-              2
-            </PaginationLink>
-          </PaginationItem>
-          <PaginationItem>
-            <PaginationLink
-              onClick={() => table.setPageIndex(2)}
-              isActive={pageIndex === 2}
-            >
-              3
-            </PaginationLink>
-          </PaginationItem>
-          <PaginationItem>
-            <PaginationEllipsis />
-          </PaginationItem>
-          <PaginationItem>
-            <PaginationNext onClick={() => null} />
-          </PaginationItem>
-        </PaginationContent>
-      </Pagination>
-    </>
-  );
-};
-
-export const ColumnVisibility = () => {
-  const table = useTable({ columns, data: invoices });
-
-  return (
-    <>
-      <div>
-        {table.getAllColumns().map((column) => (
-          <label key={column.id}>
-            <input
-              type="checkbox"
-              checked={column.isVisible()}
-              onChange={() => column.toggleVisibility()}
-            />
-            {column.id}
-          </label>
-        ))}
-      </div>
-      <Table>
-        <TableCaption>A list of your recent invoices.</TableCaption>
-        <TableHeader>
-          <TableRow>
-            {table.getAllVisibleColumns().map((column) => (
-              <TableHead key={column.id}>{column.id}</TableHead>
-            ))}
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {table.getRows().map((row) => (
-            <TableRow key={row.id}>
-              {row.getAllVisibleCells().map((cell) => (
-                <TableCell key={cell.id}>{cell.getValue()}</TableCell>
-              ))}
-            </TableRow>
-          ))}
-        </TableBody>
-        <TableFooter>
-          <TableRow>
-            <TableCell colSpan={3}>Total</TableCell>
-            <TableCell className="text-right">$2,500.00</TableCell>
-          </TableRow>
-        </TableFooter>
-      </Table>
-    </>
-  );
-};
-
-export const ColumnFiltering = () => {
-  const table = useTable({
-    columns,
-    data: invoices,
-    initialState: {
-      columnFilters: {
-        invoice: { test: true },
-      },
-    },
-  });
-
-  const columnFilters = table.getState().columnFilters;
-
-  return (
-    <>
-      <div className="mb-4">
-        Active fitlers: {JSON.stringify(columnFilters)}
-      </div>
-      <Table>
-        <TableCaption>A list of your recent invoices.</TableCaption>
-        <TableHeader>
-          <TableRow>
-            {table.getAllVisibleColumns().map((column) => (
-              <TableHead key={column.id}>
-                {column.id}
-                <input
-                  className="border"
-                  onChange={(event) =>
-                    column.setFilter({
-                      value: event.target.value,
-                      any: "",
-                      key: "",
-                      allowed: "",
-                    })
-                  }
-                />
-              </TableHead>
-            ))}
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {table.getRows().map((row) => (
-            <TableRow key={row.id}>
-              {row.getAllVisibleCells().map((cell) => (
-                <TableCell key={cell.id}>{cell.getValue()}</TableCell>
-              ))}
-            </TableRow>
-          ))}
-        </TableBody>
-        <TableFooter>
-          <TableRow>
-            <TableCell colSpan={3}>Total</TableCell>
-            <TableCell className="text-right">$2,500.00</TableCell>
-          </TableRow>
-        </TableFooter>
-      </Table>
-    </>
-  );
-};
-
-export const Sorting = () => {
-  const table = useTable({
-    columns,
-    data: invoices,
-    initialState: {
-      sorting: {
-        invoice: "asc",
-      },
-    },
-  });
-
-  return (
-    <>
-      <Table>
-        <TableCaption>A list of your recent invoices.</TableCaption>
-        <TableHeader>
-          <TableRow>
-            {table.getAllVisibleColumns().map((column) => {
-              const sorting = column.getSortingValue();
-              return (
-                <TableHead
-                  onClick={() =>
-                    column.setSort(sorting === "asc" ? "desc" : "asc")
-                  }
-                  key={column.id}
-                >
-                  {column.id}
-                  {sorting === "asc" ? "🔼" : sorting === "desc" ? "🔽" : ""}
-                </TableHead>
-              );
-            })}
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {table.getRows().map((row) => (
-            <TableRow key={row.id}>
-              {row.getAllVisibleCells().map((cell) => (
-                <TableCell key={cell.id}>{cell.getValue()}</TableCell>
-              ))}
-            </TableRow>
-          ))}
-        </TableBody>
-        <TableFooter>
-          <TableRow>
-            <TableCell colSpan={3}>Total</TableCell>
-            <TableCell className="text-right">$2,500.00</TableCell>
-          </TableRow>
-        </TableFooter>
-      </Table>
-    </>
-  );
-};
